% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{are.equidist}
\alias{are.equidist}
\title{check equidistat values}
\usage{
are.equidist(x)
}
\arguments{
\item{numeric}{vector}
}
\value{
\code{logical}
}
\description{
check if values in the array are equally spaced
}

