% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{bintext.histo1d}
\alias{bintext.histo1d}
\title{Histogram Bin Content Text}
\usage{
bintext.histo1d(histo, opt, angle = "v", round = 3, cex = 0.8, srt = 90,
  ...)
}
\arguments{
\item{\code{histo}:}{\code{histo1d} obj \cr}

\item{\code{opt}=}{"v" labels are vertical, "o": labels are orizontal}

\item{\code{round}:}{number of decimals for rounding}

\item{\code{cex}:}{font size}

\item{\code{srt}:}{angle}

\item{\code{...}:}{all the other options of the "text" function}
}
\description{
show bin content as text on top of each bin \cr
}
\details{
check also parameter \code{labels} in \code{hist}
}

