% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{Chi2.binned}
\alias{Chi2.binned}
\title{Chi2 of binned data}
\usage{
Chi2.binned(vobs, verr, vexp)
}
\arguments{
\item{hobs:}{binned observed data}

\item{herr:}{error of observed data bin}

\item{hexp:}{binned expected data}
}
\description{
Chi2 of binned data
}

