% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{Chi2.histo}
\alias{Chi2.histo}
\title{Chi2 of two histograms}
\usage{
Chi2.histo(hobs, herr, hexp)
}
\arguments{
\item{hobs:}{binned observed data}

\item{herr:}{error of observed data bin}

\item{hexp:}{binned expected data}
}
\description{
Chi2 of two histograms
}

