% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob3.R
\name{computeProb}
\alias{computeProb}
\title{compute oscillation probability}
\usage{
computeProb(alpha, beta, L, rho, E, is_anti = c(TRUE, FALSE))
}
\arguments{
\item{\code{alpha,beta}}{flavour code for the transition \code{alpha->beta}. Check \code{\link{getProb}}
for the index values. If not provided a \code{NULL} is returned.}
}
\description{
compute oscillations probabilities in the 3 flavour framework following the
linear propagation in constant density matter as derived in the paper or Barger at al.\cr
All transition probabilities are computed and can be retrieved via \code{\link{getProb}}.
the function itself returns the probability for the transition \code{alpha->beta}.\cr
This function must be called after a call to \code{\link{setMNS}} functions in order to
configure the oscillation parameters.
}

