% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob3.R
\name{computeVacuumProb}
\alias{computeVacuumProb}
\title{compute vacuum probability}
\usage{
computeVacuumProb(alpha, beta, L, E)
}
\arguments{
\item{refers}{to \code{\link{computeProb}}}
}
\description{
matrix of 3-flavour transition probabilities is filled, refer to \code{\link{computeProb}}
}

