% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.doc.R
\docType{data}
\name{data_nue_selected}
\alias{data_nue_selected}
\title{MC Neutrino Data set}
\description{
MC neutrino interaction events \cr
Variables description \cr
 event: event number \cr
 iflav: flavour \cr
 isel: selection \cr
 iint: interaction \cr
 ifv: fiducial volume \cr
 Ereco: reconstructed energy [MeV] \cr
 Etrue: true neutrino energy [MeV] \cr
 pathLen: flight path [m] \cr
 weight: event weight \cr
 IsNuE: passed NuE Selection \cr
 IsBkg: selected as background (Not NuE and Not Gamma) \cr
 IsGamma: passed Gamma control sample selection \cr
}
\author{
My Name \email{blahblah@roxygen.org}
}

