% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{dataRead.asFrame}
\alias{dataRead.asFrame}
\title{Read as data frame}
\usage{
dataRead.asFrame(file, check.firstCol = T, ...)
}
\arguments{
\item{file:}{data file name}

\item{\code{check.firstCol}:}{remove first col if all entries are \code{NA}}

\item{"...":}{inputs of \code{\link{read.delim}} othern than file name}
}
\description{
wrapper around read.delim to add the option of removing
the first column if all entries are \code{NA}
may be used to fix txt files written out with a blanck at the
beginning of each row (like files wittern via C++ cout)
}

