% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.doc.R
\docType{data}
\name{db.oscpars}
\alias{db.oscpars}
\title{Oscillation parameters data}
\description{
Oscillation Parameters. Standard 3 Flavour framework
}
\author{
My Name \email{blahblah@roxygen.org}
}

