% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.doc.R
\docType{data}
\name{db.oscpars.sterile}
\alias{db.oscpars.sterile}
\title{Oscillation parameters data}
\description{
Oscillation Parameters. Sterile neutrinos 3+1 Model framework
}
\author{
My Name \email{blahblah@roxygen.org}
}

