% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.doc.R
\docType{data}
\name{db.pars}
\alias{db.pars}
\title{Additional parameters data}
\description{
rho: earth density \cr
A0: parameter needed to calculate oscill. prob. in matter.
    (usually the actual parameter used is A=A0*rho)
}
\author{
My Name \email{blahblah@roxygen.org}
}

