% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{df.to.list}
\alias{df.to.list}
\title{data frame to by-row-list}
\usage{
df.to.list(d, level = 1)
}
\arguments{
\item{\code{d}:}{\code{data.frame} or matrix-like data structure}
}
\value{
a \code{list} of row-\code{list}s
}
\description{
convert a \code{data.frame} (or similar) into a \code{list} of rows, prior each row is converted \code{list} \cr
For default value of \code{level} (1) From an input \code{data.frame} each element of the \code{list} is a one-row \code{data.frame}.\cr
If \code{level}=2, the \code{data.frame} row itself is split into a \code{list} of elements, one of each column.
}

