% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{fun.grid.scan}
\alias{fun.grid.scan}
\title{function grid-scan}
\usage{
fun.grid.scan(grid.pars, fun = NULL, ..., .progress = "none",
  .parallel = FALSE)
}
\arguments{
\item{\code{grid.pars}:}{list of list of values to be scanned for the selected parameters \cr}

\item{\code{fun}:}{function to be evaluated. Function should be vectorized in order to evaluate the function on each point of
the grid (\code{grid.pars}) and for each element of the additional parameter (list \code{...}) passed as a vector.}

\item{\code{...}:}{other function parameters. A parameter can be passed as a vector (\code{fun} should be vectorized), in this case
an output is produced for each value of the input vactor (check the \code{return} section)}

\item{\code{progress}:}{argument of \code{alply} function in \code{plyr} package}

\item{\code{parallel}:}{run on multiple cores. Check \code{alply} function in \code{plyr} package.}
}
\value{
a \code{data.frame} with the input parameters and the corresponding function values. A separate line is produced for each
point of the \code{grid.pars} and for each value of the parameters passed as vector via the \code{...} list.
}
\description{
Perform grid-scan of the function for a subset of its parameters. Vectorizing the input function it's possible to compare the
input function shape on range of the x (y) axis for a different combination of the parameters.
}

