% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob3.R
\name{getProb}
\alias{getProb}
\title{oscillation probability}
\usage{
getProb(nu_in, nu_out)
}
\arguments{
\item{flavour}{codes are: nu_  1:e 2:mu 3:tau  -1:e_bar -2:mu_bar -3:tau_bar}
}
\description{
return oscillation probabilities for the transition nu_in -> nu_out
}
\details{
probabilities should be pre-computed via \code{\link{computeProb}}
}
\examples{
getProb(2,1): nu_mu -> nu_e \\cr
getProb(-1,-1) nu_e_bar -> nu_e_bar
}

