% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{\%on.na\%}
\alias{\%on.na\%}
\title{Replace \code{NAs}}
\usage{
a \%on.na\% b
}
\arguments{
\item{a}{is the value to be checked, b is the substitute value}
}
\description{
Replace \code{NAs}
}
\details{
Replace \code{NA} with a default value
}

