% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{\%on.nn\%}
\alias{\%on.nn\%}
\title{Replace \code{NAs} or \code{NLLs}}
\usage{
a \%on.nn\% b
}
\arguments{
\item{a}{is the value to be checked, b is the substitute value}
}
\description{
Replace \code{NAs} or \code{NLLs}
}
\details{
Replace \code{NA} or \code{NULL} with a default value
}

