% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{\%on.null\%}
\alias{\%on.null\%}
\title{Replace \code{NULLs}}
\usage{
a \%on.null\% b
}
\arguments{
\item{a}{is the value to be checked, b is the substitute value}
}
\description{
Replace \code{NULLs}
}
\details{
Replace \code{NULL} with a default value
}

