% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{grid.df}
\alias{grid.df}
\title{expand grid}
\usage{
grid.df(..., opt = NULL)
}
\arguments{
\item{\code{...}:}{set of \code{lists} to be combined (as in \code{\link{expand.grid}})}
}
\value{
a \code{data.frame}
}
\description{
just a wrapper around \code{\link{expand.grid}} for naming consistency
performs combination of all the elements in the set of lists provided via \code{...}
}

