% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StatUtils.R
\name{hist.fluct.b}
\alias{hist.fluct.b}
\title{Histogram Statistical Fluctuation Boxplot}
\usage{
\method{hist}{fluct.b}(histo, n = 1000, opt = "")
}
\value{
a boxplot-like object with desciptive
quantities for each bin.
The fivenum-style descriptive values are the one returned from stat.band function: (N.B.: they are not the usual boxplot ones):\cr
Qb,Ql,Meadia,Qr,Qe.
}
\description{
gnerate bin-by-bin statistical fluctuations calculated from \link{hist.fluct.m} function
}

