% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StatUtils.R
\name{hist.fluct.env}
\alias{hist.fluct.env}
\title{Histogram Statistical Fluctuation Envelop}
\usage{
\method{hist}{fluct.env}(histo, n = 1000, opt = "", ...)
}
\value{
points of a close polygon deliminting the envelope
}
\description{
gnerate bin-by-bin statistical fluctuations calculated from \link{hist.fluct.m} function
}
\details{
upper an lower points of the envelope are given by the value Qr and Ql defined in the \link{stat.band} function
}

