% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StatUtils.R
\name{hist.fluct.m}
\alias{hist.fluct.m}
\title{Histogram Statistical Fluctuation Matrix}
\usage{
\method{hist}{fluct.m}(histo, n = 1000, opt = "")
}
\arguments{
\item{histo:}{input histogram, statistical fluctuations are calculate around each bin count}

\item{n:}{number of repetitions}

\item{opt:}{type of fluctuation, es. "Gauss" or "Pois", check \link{stat.fluct} function for details}
}
\value{
a matrix with a colum for each histogram bin, each column as "n" values corresponding to the
"n" throws
}
\description{
generate bin-by-bin statistical fluctuations
}

