% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OscUtils.R
\name{hist.osc}
\alias{hist.osc}
\title{Histogram Reweight with Oscillation Probability}
\usage{
\method{hist}{osc}(histo, func, ..., opt = "")
}
\arguments{
\item{\code{histo}\cr}{}

\item{\code{func}}{weight function. \code{histo$mids} vector is passed as first argument}

\item{\code{...}}{additional arguments to \code{func}}

\item{opt:}{other configuration options}
}
\value{
weighted histogram. \code{histo} obj contains a vector with the used wieghts
}
\description{
weight histogram counts with value of function \code{func} evaluated at the center of the bins (i.e.:\code{mids} vector in \link{hist})
}

