% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{histo1d.errs}
\alias{histo1d.errs}
\title{Compute  1D histogram bin erros}
\usage{
histo1d.errs(x, opt = NULL)
}
\arguments{
\item{vector}{of bin contents}
}
\value{
vector of bin errors
}
\description{
compute error of each bin content, default is to take the \code{sqrt} of the
bin content
}

