% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{histo1d}
\alias{histo1d}
\title{1D Histogram}
\usage{
histo1d(x = NULL, weight = NULL, breaks, na.rm = TRUE, show = TRUE,
  main, xlab, ylab, ...)
}
\arguments{
\item{\code{x}:}{list of entries}

\item{\code{w}:}{list of weights}

\item{\code{breaks}:}{bins edges}

\item{\code{show}:}{plot the histo}

\item{\code{main}:}{histo title}

\item{\code{xlab,ylab}:}{axis titles}

\item{\code{...}:}{other graphic options}
}
\value{
obj elements:\cr

\code{nentries}

\code{outlayers}

\code{counts}

\code{breaks}

\code{nbins}

\code{binw}: bin width

\code{errs}: bin error

\code{equidist}

\code{main,xlab,ylab}

\code{opts}: arguments in \code{...}

\code{call}
}
\description{
General 1D (weighted) histogram structure
}

