% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{histo2d.errs}
\alias{histo2d.errs}
\title{Compute  2D histogram bin erros}
\usage{
histo2d.errs(x, opt = NULL)
}
\arguments{
\item{matrix}{of bin contents}
}
\value{
matrix of bin errors
}
\description{
compute error of each bin content, default is to take the \code{sqrt} of the
bin content
}

