% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{histo2d}
\alias{histo2d}
\title{2D Histogram}
\usage{
histo2d(x, y = NULL, weight = NULL, breaks.x, breaks.y, na.rm = TRUE,
  main, xlab, ylab, show = TRUE, legend = TRUE, col = c("white",
  heat.colors(12)), ...)
}
\arguments{
\item{\code{x}:}{list of entries}

\item{\code{w}:}{list of weights}

\item{\code{breaks}:}{bins edges}

\item{\code{show}:}{plot the histo}

\item{\code{main}:}{histo title}

\item{\code{xlab,ylab}:}{axis titles}

\item{\code{col}}{}

\item{\code{...}:}{other graphic options}
}
\value{
obj elements:\cr

\code{nentries}

\code{outlayers}

\code{counts}

\code{breaks}

\code{nbins}

\code{binw}: bin width

\code{errs}: bin error

\code{equidist}

\code{main,xlab,ylab}

\code{opts}: arguments in \code{...}

\code{call}
}
\description{
General 2D (weighted) histogram structure
}

