% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{is.compiled}
\alias{is.compiled}
\title{Check if function is compiled}
\usage{
is.compiled(func)
}
\arguments{
\item{\code{func}}{function to check}
}
\value{
\code{logical}
}
\description{
check if the function hase been byte-compiled (Returns \code{TRUE} if it finds the text "bytecode:")
}

