% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{list.to.df}
\alias{list.to.df}
\title{from a \code{list} of rows to a \code{\link{data.frame}}}
\usage{
list.to.df(list)
}
\value{
a \code{data.frame}
}
\description{
each element of the input \code{list} can be one-row \code{data.frame}
or a \code{list} of element each representing a column of the final \code{data.frame}
}

