% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{LogLike.binned}
\alias{LogLike.binned}
\title{LogLike of binned data}
\usage{
LogLike.binned(vobs, vexp, opt = "")
}
\arguments{
\item{hobs:}{binned observed data}

\item{hexp:}{binned expected data}
}
\description{
compute negative loglike
}

