% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{LogLike.histo}
\alias{LogLike.histo}
\title{LogLike of two histogram}
\usage{
LogLike.histo(hobs, hexp, opt = "")
}
\arguments{
\item{hobs:}{binned observed data}

\item{hexp:}{binned expected data}
}
\description{
compute negative loglike
}

