% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{logspaced}
\alias{logspaced}
\title{log10-spaced values}
\usage{
logspaced(n, min, max, force.first = F)
}
\arguments{
\item{\code{n}:}{number of points}

\item{\code{min,max}:}{interval range}

\item{\code{force.first}:}{put first element equal to \code{min} (last element is always equal to \code{max})}
}
\description{
compute a vector of \code{n} log10-spaced values in a given range
}
\details{
used when you need finer sampling of functions which are varying more rapidly
in the lower part of the sampling range
}

