% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{lshape.hist1d}
\alias{lshape.hist1d}
\title{1D Histo line shape}
\usage{
lshape.hist1d(histo1d, opt = NULL)
}
\arguments{
\item{\code{histo1d}}{}
}
\value{
list with x and y vectors defining the hist shape
}
\description{
calculate line defining the shape of the histogram
}

