% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{mid.points}
\alias{mid.points}
\title{compute array mid-points}
\usage{
mid.points(x)
}
\arguments{
\item{\code{x}}{\code{numeric} list of values}
}
\value{
list of n-1 elements
}
\description{
compute n-1 mid-points of an array of n elements
}

