% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob3.R
\docType{data}
\name{my.env}
\alias{my.env}
\title{environment for global variables}
\format{\preformatted{<environment: 0x7fe6e4d63d50> 
}}
\usage{
my.env
}
\description{
main functions \code{\link{setMNS}},\code{\link{computeProb}},\code{\link{computeVaccumProb}}
}
\keyword{datasets}

