% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{new.win}
\alias{new.win}
\title{new window}
\usage{
new.win(title, opt = "")
}
\arguments{
\item{\code{title}:}{window title}
}
\description{
open a new window for graphics.
For OSX it calls quartz function
}

