% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NuUtils.R
\name{Nu.Ereco}
\alias{Nu.Ereco}
\title{Nu Reconstructed Energy}
\usage{
Nu.Ereco(Mlepton, Plepton, cos.theta)
}
\arguments{
\item{Mlepton:}{Mass of the outgoing lepton [MeV] \cr}

\item{Plepton:}{Momentum of the outgoing lepton  [MeV/c] \cr}

\item{cos.theta:}{theta = angle of the outgoing lepton  wrt to the incoming neutrino \cr}
}
\description{
Reconstructed Neutrino Energy based on quasielastic kinematic
}

