% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NuUtils.R
\name{Nu.ErecoAsEle}
\alias{Nu.ErecoAsEle}
\title{NuE Reconstructed Energy}
\usage{
Nu.ErecoAsEle(Plepton, cos.theta)
}
\arguments{
\item{Elepton:}{Energy of the outgoing lepton (electron) [MeV]\cr}

\item{Plepton:}{Momentum of the outgoing lepton (electron) [MeV/c]\cr}

\item{cos.theta:}{theta = angle of the outgoing lepton (electron) wrt to the incoming neutrino\cr}
}
\description{
Reconstructed Neutrino Energy based on quasielastic kinematic
electron neutrino hyp.
}

