% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NuUtils.R
\name{Nu.ErecoAsMuon}
\alias{Nu.ErecoAsMuon}
\title{NuMu Reconstructed Energy}
\usage{
Nu.ErecoAsMuon(Plepton, cos.theta)
}
\arguments{
\item{Elepton:}{Energy of the outgoing lepton (muon) [MeV] \cr}

\item{Plepton:}{Momentum of the outgoing lepton (muon) [MeV]\cr}

\item{cos.theta:}{theta = angle of the outgoing lepton (muon) wrt to the incoming neutrino\cr}
}
\description{
Reconstructed Neutrino Energy based on quasielastic kinematic
muon neutrino hyp.
}

