% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OscUtils.R
\name{nu.trans.exp}
\alias{nu.trans.exp}
\title{transition \code{plotmath} expression}
\usage{
nu.trans.exp(nu_in, nu_out)
}
\arguments{
\item{Standard}{flavour codes of the initial and final states  (e:1,mu:2,tau:3; negative for antiparticles)}
}
\description{
quote the \code{plotmath} expression for the neutrino oscillation transition \code{nu_in} -> \code{nu_out}
}

