% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{plot.histo1d}
\alias{plot.histo1d}
\title{plot \code{histo1d}}
\usage{
\method{plot}{histo1d}(histo1d, main, xlab, ylab, ...)
}
\arguments{
\item{\code{histo1d}:}{a histo1d obj}

\item{\code{main,xlab,ylab}:}{arguments of \code{plot}}

\item{\code{...}:}{other graphics options}
}
\description{
Plot \code{histo1d} by means of the \code{plot.hist} function
}

