% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{plot.histo2d}
\alias{plot.histo2d}
\title{plot \code{histo2d}}
\usage{
\method{plot}{histo2d}(histo2d, legend = T, col = c("white",
  heat.colors(12)), main, xlab, ylab, ...)
}
\arguments{
\item{\code{histo2d}:}{a histo2d obj}

\item{\code{legend}:}{add color gradied scale for z axis (if \code{TRUE} \code{image.plot} is used)}

\item{\code{col}:}{heat color map for the \code{legend}}

\item{\code{xlab,ylab}:}{axis titles}

\item{\code{...}:}{graphics options}
}
\description{
Plot \code{histo2d} by means of the \code{image} function
}

