% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OscUtils.R
\name{prob.env}
\alias{prob.env}
\title{Oscillation Probability Envelope}
\usage{
prob.env(e, func, ...)
}
\arguments{
\item{e:}{vector or energy points}

\item{func:}{oscillation probability function}

\item{...:}{list of arguments to func}
}
\value{
a list containing \code{x} and \code{y} coordinates of the envelope polygon
}
\description{
calculate for a given range of values of one of the oscillation parameters the
envelope containing the max and min oscillation probability, as a function of the energy
}

