% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OscProbApprox.R
\name{ProbNuMuToNuE_cmp}
\alias{ProbNuMuToNuE_cmp}
\title{NuEAppearance compiled
Check ProbNuMuToNuE for reference
Temporary keeping both}
\usage{
ProbNuMuToNuE_cmp(E, L, s2_12, s2_13, s2_23, dm2_21, dm2_32, MH, dcp)
}
\description{
NuEAppearance compiled
Check ProbNuMuToNuE for reference
Temporary keeping both
}

