% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OscProbApprox.R
\name{ProbNuMuToNuMu}
\alias{ProbNuMuToNuMu}
\title{NuMuDisappearance}
\usage{
ProbNuMuToNuMu(E, L, s2_12, s2_13, s2_23, dm2_21, dm2_32, MH, dcp)
}
\arguments{
\item{E,L}{neutrino Energy and Baseline. Units: [GeV] and [km] or [MeV] and [m], ...}

\item{s2_12:}{sin^{2}_{12}}

\item{s2_13:}{sin^{2}_{13}}

\item{s2_23:}{sin^{2}_{23}}

\item{dm2_21:}{Delta_m^2_{21} [eV^2]}

\item{dm2_32:}{Delta_m^2_{32} [eV^2]}

\item{MH:}{Mass Hierachy, MH=1 for Normal Hierachy NH, MH=-1 for Inverted Hierarchy IH}

\item{dcp:}{delta CP phase}
}
\value{
Probability
}
\description{
NuMuDisappearance
}
\details{
3 Flavour approximate formulta for NuMu->NuMu
}

