% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OscProbApprox.R
\name{ProbNuMuToNuMu_cmp}
\alias{ProbNuMuToNuMu_cmp}
\title{NuMuDisappearance compiled
Check ProbNuMuToNuMu for reference
Temporary keeping both}
\usage{
ProbNuMuToNuMu_cmp(E, L, s2_12, s2_13, s2_23, dm2_21, dm2_32, MH, dcp)
}
\description{
NuMuDisappearance compiled
Check ProbNuMuToNuMu for reference
Temporary keeping both
}

