% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{rm.duplnames}
\alias{rm.duplnames}
\title{remove duplicated list-element names}
\usage{
rm.duplnames(list)
}
\arguments{
\item{\code{list}}{}
}
\value{
\code{list} with removed duplicates
}
\description{
remove duplicates element names in a list: only first occurrence is retained
}

