% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob3.R
\name{setAndComputeProb}
\alias{setAndComputeProb}
\title{set and compute ascill. probability nu_in -> nu_out}
\usage{
setAndComputeProb(sin2_12, sin2_13, sin2_23, dm2_21, dm2_23,
  MH = factor(c("NH", "IH")), delta, is_anti = c(T, F), L, rho, E, nu_in,
  nu_out)
}
\description{
wrapper function to setup and compute transition probabilities, it calls in sequence \code{\link{setMNS}}) and \code{\link{computeProb}})
}
\details{
Assumes \code{sin} are passed as \code{sin^2} (i.e.: \code{squared=T} in \code{\link{setMNS}})\cr
\code{MH} is of type \code{factor}\cr
}

