% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prob3.R
\name{setMNS}
\alias{setMNS}
\title{set MNS matrix, mass splittings etc.}
\usage{
setMNS(x12, x13, x23, dm21, dmAtm, delta, kSquared = c(TRUE, FALSE),
  is_anti = c(TRUE, FALSE))
}
\arguments{
\item{Specify}{the neutrino oscillation parameters, and energy,
the final boolean specifies which form  of mixing angle is input \cr
          \code{x12   ,  x13   ,  x23   ,  dm21  ,  dmAtm  ,  d_cp, kSquared , is_anti}\cr
Notation: m21=mSol \cr
If \code{dmAtm>0} Normal Hierachy is assumed (NH) and \code{dm32 = dmAtm (and dm31=dm21+dm32)}\cr
If \code{dmAtm<0} Inverted Hierarchy is assumed (IH), that is \code{dmAtm=dm31<0 and dm32 = -abs(dmAtm)-dm21 <0 }\cr
By default the "One dominant mass scale" approx is used, therefore only \code{dm21} and \code{dm32} are used\cr
\code{kSquared=T: sin^2(x) F: sin^2(2x)} \cr
\code{is_anti=T} if for anti-neutrino propagation \cr}
}
\value{
Following global objects are filled: \code{mix} (MNS matrix), \code{dm,dmVacVac} (mass matrix).
}
\description{
determine the neutrino oscillation parameters
This function must be called before propagation functions
\code{\link{getProb}}, \code{\link{getVaccumProb}}
}
\details{
Computation follows the derivation of Barger et al. \code{Phys. Rev. D22(1980) 2718.}, code is adapted form its original C-version
available here \code{www.phy.duke.edu/~raw22/public/Prob3++}
}

