% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StatUtils.R
\name{stat.band}
\alias{stat.band}
\title{List Statistical Descriptive Quantities}
\usage{
stat.band(x, quant = 0.68)
}
\arguments{
\item{x:}{input array of values \cr
quant: "quantile around the median", see return values for a description.}
}
\value{
Mean, Median,SD (standard deviation), Min, Max \cr
Qr/Ql: value which limit quant/2 portion of the sample on the right/left of the median \cr
it means that in the range [Qr,Ql] it is contained quant fraction of the sample \cr
while each of the ranges [Ql,Median] and [Median,Qr] contains quant/2 fraction of the sample\cr
Qb/Qe: 5\% and 95\% quantile
}
\description{
Stat utils
}
\details{
example stopifnot(!missing(x), is.matrix(x), dim(x)==c(2,2), x>0)
}

