% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StatUtils.R
\name{stat.fluct}
\alias{stat.fluct}
\title{Statistical Fluctuation}
\usage{
stat.fluct(n, opt = "Pois")
}
\arguments{
\item{value}{to be flucuated}

\item{option}{("Poisson", "Gauss")}
}
\value{
fluctuated value
}
\description{
apply a Poisson (default) or Gaussian Statistical Fluctuation \cr
to the input number n.\cr
For Gauss the variation is sampled from a gaussian with sigma=sqrt(n) and mean 0\cr
For Poisson the variation is sampled from a poissonian with lambda=sqrt(n)
}

