% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OscProbSterile.R
\name{Sterile.ProbNuEToNuE}
\alias{Sterile.ProbNuEToNuE}
\title{NuE Disppearance via Sterile Transition}
\usage{
Sterile.ProbNuEToNuE(E, L, s2_2ee, dm2_41)
}
\arguments{
\item{E,L:}{neutrino Energy and Baseline. Units: [GeV] and [km] or [MeV] and [m], ...}

\item{s2_ee:}{sin^{2}_{ee}}

\item{dm2_41:}{in  Delta_m^2_{41} [eV^2]}
}
\value{
probability
}
\description{
NuE Disppearance via Sterile Transition
}
\details{
NuE disappearance
in the Model 3+1.\cr
Approximate expression
}

