% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hist.R
\name{to.hist}
\alias{to.hist}
\title{convert \code{histo1d} to \code{histogram}}
\usage{
to.hist(histo1d)
}
\value{
a \code{hist} obj
}
\description{
build a \code{histogram} obj from a \code{histo1d} one
}

